requirements.txt
flask 
torch 
torchvision 
torchaudio 
opencv-python 
numpy 
openai ==0.28.0
python-dotenv 
pillow
pandas
requests
